quickSort tests added
