+++ algorithms/quickSort/quickSort.js
@@ -1,9 +1,4 @@
-/* eslint-disable */
-// TODO: enable linting
-
-const exampleList = [4, 56, 1, 14, 11, 67, 23, 3, 7, 89, 16, 2, 91, 34];
-
-const quickSort = (listArg) => {
+function quickSort(listArg) {
   let list = listArg;
   if (list.length < 2) return list;
   const leftList = [];
@@ -15,13 +10,12 @@ const quickSort = (listArg) => {
   //   item < pivotValue ? leftList.push(item) : rightList.push(item);
   // }
   list.forEach((item) => {
+    // eslint-disable-next-line no-unused-expressions
     item < pivotValue ? leftList.push(item) : rightList.push(item);
   });
 
   const sortedList = quickSort(leftList).concat([pivotValue], quickSort(rightList));
   return sortedList;
-};
+}
 
-module.exports.runQuickSort = function () {
-  console.log(quickSort(exampleList));
-};
+module.exports = quickSort;
