+++ algorithms/quickSort/quickSortNotes.md
@@ -1,0 +1,1 @@
+quickSort Notes go here!
