+++ algorithms/quickSort/quickSortSpec.md
@@ -1,0 +1,1 @@
+quickSort Spec goes here!
