+++ algorithms/selectionSort1/selectionSort1.js
